Refactor CollectionHelperTest
Some classes don't need to be testet, eg `singletonList(..).getClass()`, since
those can't be used as a type. Also, checking for the exact type is not
appropriate. Instead, it must be checked if the created collection can be
assigned to the required type.

CustomSet also appears faulty. It had no public default constructor, so a
CustomSet can never be created and `newCollection` would have to fail. Until
now, a hash set was intended as the expected result, but this would only
lead to runtime errors.

Signed-off-by: Yannick da Silva Bröker <ybroeker@techfak.uni-bielefeld.de>